split out grid calculations; calc colors for grid
ripped the relevant code out of hq_candidates.ipynb to make grid_calculations.ipynb. This also updates the bulla_2019_rates.csv to include colors for the peaks in all filters. I also added the calculations for the ztfi filter of the transient sample in gp_calculations for completeness, but the paucity of zffi observations means very few are actually present. The plots for color can now be made in gp_calculationsipynb. May be advantageous to rip out all the plotting material into a separate file since gp_calculations is getting a bit long